dynamic padding based on header height
this was an interesting experiment. a vue component measures its own height and sets a data property, which is available on export. then, the parent component reads this height by ref, then computes inline style based on that height. i will invariably end up using a fixed height, but this was an interesting detour